Make `$field->toPercentString()` work with locales that use a comma as decimal point.
Some locales are using a comma instead of a dot as decimal point. When setting PHP’s locale to e.g. German, floats converted to strings also will be output with a comma, causing problems when using them as CSS values. My change replaces the float conversion with a more abstract solution, that should work with any decimal point and makes it work with any locale.